requirements.txt
Django==4.1.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
requests~=2.28.1
uuid==1.30